requirements.txt
tqdm
scikit-learn
pillow
pyyaml
tabulate
matplotlib
pandas
opencv-python
torchvision
torch
timm>=0.9.10
huggingface-hub>=0.23.0
open_clip_torch
einops
imageio
ipympl
numpy
tensorboard
tensorboardX
blobfile>=1.0.5
natsort